+++ flutterannotations_generator/example/lib/model/home_banner_model.g.dart
@@ -1,31 +1,31 @@
-// GENERATED CODE - DO NOT MODIFY BY HAND
-
-part of 'home_banner_model.dart';
-
-// **************************************************************************
-// JsonSerializableGenerator
-// **************************************************************************
-
-HomeBannerModel _$HomeBannerModelFromJson(Map<String, dynamic> json) {
-  return HomeBannerModel(
-      json['desc'] as String,
-      json['id'] as int,
-      json['imagePath'] as String,
-      json['isVisible'] as int,
-      json['order'] as int,
-      json['title'] as String,
-      json['type'] as int,
-      json['url'] as String);
-}
-
-Map<String, dynamic> _$HomeBannerModelToJson(HomeBannerModel instance) =>
-    <String, dynamic>{
-      'desc': instance.desc,
-      'id': instance.id,
-      'imagePath': instance.imagePath,
-      'isVisible': instance.isVisible,
-      'order': instance.order,
-      'title': instance.title,
-      'type': instance.type,
-      'url': instance.url
-    };
+//// GENERATED CODE - DO NOT MODIFY BY HAND
+//
+//part of 'home_banner_model.dart';
+//
+//// **************************************************************************
+//// JsonSerializableGenerator
+//// **************************************************************************
+//
+//HomeBannerModel _$HomeBannerModelFromJson(Map<String, dynamic> json) {
+//  return HomeBannerModel(
+//      json['desc'] as String,
+//      json['id'] as int,
+//      json['imagePath'] as String,
+//      json['isVisible'] as int,
+//      json['order'] as int,
+//      json['title'] as String,
+//      json['type'] as int,
+//      json['url'] as String);
+//}
+//
+//Map<String, dynamic> _$HomeBannerModelToJson(HomeBannerModel instance) =>
+//    <String, dynamic>{
+//      'desc': instance.desc,
+//      'id': instance.id,
+//      'imagePath': instance.imagePath,
+//      'isVisible': instance.isVisible,
+//      'order': instance.order,
+//      'title': instance.title,
+//      'type': instance.type,
+//      'url': instance.url
+//    };