+++ flutterannotations/CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## 1.0.2
+
+* fix the parameter of 'failCallback' 
+* fix  the ' DioError [DioErrorType.RESPONSE]: data parsing exception...' bug
+
 ## 1.0.1
 
 * add interceptors in 'Rest' annotation